+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - chore(accelerator): update sdk v3
 - chore(active-directory): update sdk v3
+- chore(audit-manager): update sdk v3
 
 ## [1.12.0] - 2025-04-02